chore: rename cosem to cellmap
