+++ src/components/cellmapLogo.png
