+++ src/components/segmentation_gradient.png
