+++ src/components/segmentation_gradient_full.png
